Create error handling method

+++ products-frontend/src/app/components/product/product-delete/product-delete.component.ts
@@ -5,6 +5,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { FormsModule } from '@angular/forms';
 import { MatCardModule } from '@angular/material/card';
 import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
 
 @Component({
   selector: 'app-product-delete',
@@ -12,13 +13,14 @@ import { MatFormFieldModule } from '@angular/material/form-field';
   imports: [
     FormsModule,
     MatCardModule,
-    MatFormFieldModule
+    MatFormFieldModule,
+    MatInputModule
   ],
   templateUrl: './product-delete.component.html',
   styleUrl: './product-delete.component.css'
 })
 export class ProductDeleteComponent {
-  product: Product = { id: 0, name: '', price: 0 };
+  product: Product = { id: '0', name: '', price: 0 };
 
   constructor(
     private productService: ProductService,
@@ -27,21 +29,36 @@ export class ProductDeleteComponent {
   ) {}
 
   ngOnInit(): void {
-    const id = +(this.route.snapshot.paramMap.get("id") ?? 0);
-    this.productService.readById(id).subscribe((product) => {
-      this.product = product;
-    });
-  }
-
-  deleteProduct(): void {
-    if (this.product.id !== undefined) {
-      this.productService.delete(this.product.id).subscribe(() => {
-        this.productService.showMessage("Produto excluido com sucesso!");
-        this.router.navigate(["/products"]);
+    // Obtém o ID da rota como string
+    const id = this.route.snapshot.paramMap.get("id");
+  
+    if (id) {
+      this.productService.readById(id).subscribe({
+        next: (product) => {
+          this.product = product;
+        },
+        error: (error) => {
+          this.productService.showMessage("Produto não encontrado!");
+          this.router.navigate(["/products"]);
+        }
       });
+    } else {
+      this.productService.showMessage("ID do produto é inválido.");
+      this.router.navigate(["/products"]);
     }
   }
-
+      
+  deleteProduct(): void {
+    if (this.product && this.product.id !== undefined) {
+      this.productService.delete(this.product.id).subscribe(() => {
+        this.productService.showMessage("Produto excluído com sucesso!");
+        this.router.navigate(["/products"]);
+      });
+    } else {
+      this.productService.showMessage("Produto inválido para exclusão.");
+    }
+  }
+  
   cancel(): void {
     this.router.navigate(["/products"]);
   }
